cartview修正

+++ db/migrate/20230218082910_change_data_quantity_to_cart_items.rb
@@ -1,5 +1,5 @@
 class ChangeDataQuantityToCartItems < ActiveRecord::Migration[7.0]
   def change
-    change_column :cart_items, :quantity, :integer
+    change_column :cart_items, :quantity, :float
   end
 end
